Don't shut down process pool if it never started.

+++ firebase/__init__.py
@@ -1,6 +1,6 @@
 import atexit
 
-from ._async import process_pool
+from ._async import process_pool, _process_pool
 from .firebase import *
 
 
@@ -10,6 +10,8 @@ def close_process_pool():
     Clean up function that closes and terminates the process pool
     defined in the ``async`` file.
     """
-    process_pool.close()
-    process_pool.join()
-    process_pool.terminate()
+    # if the pool doesn't already exist, don't create it just to stop it.
+    if _process_pool is not None:
+        process_pool.close()
+        process_pool.join()
+        process_pool.terminate()
